Handle multiple submit buttons
When using bs-confirmation with multiple input[type=submit] buttons, the current behaviour is to trigger form.submit after confirmation is received. If you have multiple submit buttons (with unique names/values) then this submits the form with the wrong action. This commit changes that behaviour so that if there are multiple submit buttons, the confirmation is destroyed then the click event reinvoked causing the original submit action to occur (which is browser submitting the form with the correct name/value specified for the clicked button)